@@ -19,19 +19,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^17.3.9",
-    "@angular/animations": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.5",
+    "@angular/animations": "^18.2.5",
     "@angular/cdk": "^16.2.9",
-    "@angular/cli": "^17.3.9",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/language-service": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/cli": "^18.2.5",
+    "@angular/common": "^18.2.5",
+    "@angular/compiler": "^18.2.5",
+    "@angular/compiler-cli": "^18.2.5",
+    "@angular/core": "^18.2.5",
+    "@angular/forms": "^18.2.5",
+    "@angular/language-service": "^18.2.5",
+    "@angular/platform-browser": "^18.2.5",
+    "@angular/platform-browser-dynamic": "^18.2.5",
+    "@angular/router": "^18.2.5",
     "@coreui/angular": "~4.5.27",
     "@coreui/angular-chartjs": "~4.5.27",
     "@coreui/chartjs": "^3.1.2",
@@ -55,7 +55,7 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular/localize": "^17.3.12",
+    "@angular/localize": "^18.2.5",
     "@types/jasmine": "^5.1.1",
     "@types/lodash-es": "^4.17.10",
     "@types/node": "^18.18.6",
